+++ src/compas_slicer/utilities/utils.py
@@ -402,7 +402,7 @@ def get_mesh_cotmatrix_igl(mesh, fix_boundaries=True):
     :class: 'scipy.sparse.csr_matrix'
         sparse matrix (dimensions: #V x #V), laplace operator, each row i corresponding to v(i, :)
     """
-    check_package_is_installed('igl')
+    # check_package_is_installed('igl')
     import igl
     v, f = mesh.to_vertices_and_faces()
     C = igl.cotmatrix(np.array(v), np.array(f))
@@ -431,7 +431,7 @@ def get_mesh_cotans_igl(mesh):
     :class: 'np.array'
         Dimensions: F by 3 list of 1/2*cotangents corresponding angles
     """
-    check_package_is_installed('igl')
+    # check_package_is_installed('igl')
     import igl
     v, f = mesh.to_vertices_and_faces()
     return igl.cotmatrix_entries(np.array(v), np.array(f))